+++ plasmapy/formulary/parameters.py
@@ -3,7 +3,6 @@
 __all__ = [
     "Alfven_speed",
     "Bohm_diffusion",
-    "Debye_number",
     "Hall_parameter",
     "ion_sound_speed",
     "kappa_thermal_speed",
@@ -18,7 +17,6 @@ __aliases__ = [
     "betaH_",
     "cs_",
     "DB_",
-    "nD_",
     "pmag_",
     "pth_",
     "rho_",
@@ -65,13 +63,8 @@ eps0_si_unitless = eps0.value
 k_B_si_unitless = k_B.value
 
 funcs_to_deprecate_wrap = [  # (module_name, func_name)
-    ("lengths", "Debye_length"),
-    ("lengths", "lambdaD_"),
-    ("lengths", "gyroradius"),
-    ("lengths", "rc_"),
-    ("lengths", "rhoc_"),
-    ("lengths", "inertial_length"),
-    ("lengths", "cwp_"),
+    ("dimensionless", "Debye_number"),
+    ("dimensionless", "nD_"),
     ("frequencies", "gyrofrequency"),
     ("frequencies", "oc_"),
     ("frequencies", "wc_"),
@@ -82,6 +75,13 @@ funcs_to_deprecate_wrap = [  # (module_name, func_name)
     ("frequencies", "wlh_"),
     ("frequencies", "upper_hybrid_frequency"),
     ("frequencies", "wuh_"),
+    ("lengths", "Debye_length"),
+    ("lengths", "lambdaD_"),
+    ("lengths", "gyroradius"),
+    ("lengths", "rc_"),
+    ("lengths", "rhoc_"),
+    ("lengths", "inertial_length"),
+    ("lengths", "cwp_"),
 ]
 for modname, name in funcs_to_deprecate_wrap:
     globals()[name] = deprecated(
@@ -1196,80 +1196,6 @@ def Hall_parameter(
 
 betaH_ = Hall_parameter
 """Alias to `~plasmapy.formulary.parameters.Hall_parameter`."""
-
-
-@validate_quantities(
-    T_e={"can_be_negative": False, "equivalencies": u.temperature_energy()},
-    n_e={"can_be_negative": False},
-)
-def Debye_number(T_e: u.K, n_e: u.m ** -3) -> u.dimensionless_unscaled:
-    r"""Return the number of electrons within a sphere with a radius
-    of the Debye length.
-
-    **Aliases:** `nD_`
-
-    Parameters
-    ----------
-    T_e : `~astropy.units.Quantity`
-        Electron temperature.
-
-    n_e : `~astropy.units.Quantity`
-        Electron number density.
-
-    Raises
-    ------
-    `TypeError`
-        If either argument is not a `~astropy.units.Quantity`.
-
-    `astropy.units.UnitConversionError`
-        If either argument is in incorrect units.
-
-    `ValueError`
-        If either argument contains invalid values.
-
-    Warns
-    -----
-    : `~astropy.units.UnitsWarning`
-        If units are not provided, SI units are assumed.
-
-    Returns
-    -------
-    N_D : `~astropy.units.Quantity`
-        Number of electrons within a sphere with a radius of the Debye
-        length.
-
-    Notes
-    -----
-    The Debye number is the number of electrons contained within a
-    sphere with a radius of a Debye length and is given by
-
-    .. math::
-        N_D = \frac{4π}{3} n_e λ_D^3
-
-    The Debye number is also known as the plasma parameter.
-
-    Collective behavior requires :math:`N_D ≫ 1`\ .
-
-    See Also
-    --------
-    ~plasmapy.formulary.lengths.Debye_length
-
-    Examples
-    --------
-    >>> from astropy import units as u
-    >>> Debye_number(5e6*u.K, 5e9*u.cm**-3)
-    <Quantity 2.17658...e+08>
-
-    """
-
-    lambda_D = lengths.Debye_length(T_e, n_e)
-    N_D = (4 / 3) * np.pi * n_e * lambda_D ** 3
-
-    return N_D
-
-
-nD_ = Debye_number
-"""Alias to `~plasmapy.formulary.parameters.Debye_number`."""
 
 
 @validate_quantities
